+++ plugins/product-group-detergent/src/lib.rs
@@ -6,8 +6,8 @@
 
 use dpp_plugin_sdk::export_plugin;
 use dpp_plugin_sdk::traits::{
-    DppProductGroupPlugin, METRIC_CO2E_SCORE, PluginComplianceStatus, PluginError, PluginFieldError,
-    PluginIdentity, PluginInput, PluginResult, SchemaVersionRange,
+    DppProductGroupPlugin, METRIC_CO2E_SCORE, PluginComplianceStatus, PluginError,
+    PluginFieldError, PluginIdentity, PluginInput, PluginResult, SchemaVersionRange,
 };
 use dpp_plugin_sdk::validate::{Validator, num};
 use serde_json::Value;
@@ -28,13 +28,13 @@ impl DppProductGroupPlugin for DetergentPlugin {
     fn schema_version_range(&self) -> SchemaVersionRange {
         SchemaVersionRange {
             min_version: "1.0.0".into(),
-            max_version: "1.1.0".into(),
+            max_version: "1.2.0".into(),
         }
     }
 
     fn validate_input(&self, input: &PluginInput) -> Result<(), PluginError> {
         Validator::new(input)
-            .require_gtin("gtin")
+            .require_product_identifier("productIdentifier")
             .require_str("productType")
             .require_str("format")
             .require_non_empty_array("surfactants")
@@ -98,7 +98,7 @@ mod tests {
 
     fn valid() -> Value {
         json!({
-            "gtin": "12345678901231",
+            "productIdentifier": {"scheme": "gs1", "gtin": "12345678901231"},
             "productType": "laundry",
             "format": "liquid",
             "surfactants": [